alx_travel_app/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11.2,<0.12
mysqlclient>=2.2.4,<3.0
django-cors-headers>=4.3.1,<5.0
drf-yasg>=1.21.7,<2.0
celery>=5.4.0,<6.0
